tutorial/requirements.txt
https://www.djangoproject.com/download/1.7.b4/tarball/
-e git+https://github.com/django/django-contrib-comments.git#egg=comments
django-jsonfield==0.9.12
